+++ api/lib/mno_enterprise/api.rb
@@ -1,5 +1,6 @@
 require 'action_view' # To fix "uninitialized constant Haml::ActionView"
 require 'jbuilder'
+require 'coffee-rails'
 require 'haml'
 require 'credit_card_validations'