+++ core/app/models/mno_enterprise/tenant_reporting.rb
@@ -1,0 +1,10 @@
+module MnoEnterprise
+  class TenantReporting < BaseResource
+    # == Class Methods ========================================================
+
+    # This is a singleton resource
+    def self.table_name
+      'tenant_reporting'
+    end
+  end
+end